feat: Refactoring functions #1
