+++ .vscodeignore
@@ -11,3 +11,4 @@ vsc-extension-quickstart.md
 **/.eslintrc.json
 **/*.map
 **/*.ts
+forTest/**
